requirements.txt
kafka-python
pymongo<4.0